Personal Cybersecurity Website
This website highlights my interest and passion in cybersecurity. My resume and experience are provided on this website for potential employers. I also share my previous and current projects that I am working on and the mission behind them. My blog is a place where I share personal hackathon experiences, step by step instructions on hacking projects, and my life as a women in information security. This is to inspire women as well as men aspiring to be a cyber professional and what the daily life of one entails.